move functions to stores
